refactor(wasm): asbtract modules
